+++ sdk/appconfiguration/app-configuration/src/models.ts
@@ -293,9 +293,24 @@ export interface ListSettingsOptions extends OptionalFields {
 export interface ListConfigurationSettingsOptions extends OperationOptions, ListSettingsOptions {}
 
 /**
+ * @interface
+ * An interface that tracks the settings for paged iteration
+ */
+export interface PageSettings {
+  /**
+   * @member {string} [continuationToken] The token that keeps track of where to continue the iterator
+   */
+  continuationToken?: string;
+  // The appconfig service doesn't currently support letting you select a page size
+  // so we're ignoring their setting for now.
+}
+
+/**
  * A page of configuration settings and the corresponding HTTP response
  */
-export interface ListConfigurationSettingPage extends HttpResponseField<SyncTokenHeaderField> {
+export interface ListConfigurationSettingPage
+  extends HttpResponseField<SyncTokenHeaderField>,
+    PageSettings {
   /**
    * The configuration settings for this page of results.
    */
@@ -312,7 +327,7 @@ export interface ListRevisionsOptions extends OperationOptions, ListSettingsOpti
 /**
  * A page of configuration settings and the corresponding HTTP response
  */
-export interface ListRevisionsPage extends HttpResponseField<SyncTokenHeaderField> {
+export interface ListRevisionsPage extends HttpResponseField<SyncTokenHeaderField>, PageSettings {
   /**
    * The configuration settings for this page of results.
    */
@@ -331,3 +346,18 @@ export interface SetReadOnlyResponse
   extends ConfigurationSetting,
     SyncTokenHeaderField,
     HttpResponseField<SyncTokenHeaderField> {}
+
+/**
+ * Options that control how to retry failed requests.
+ */
+export interface RetryOptions {
+  /**
+   * The maximum number of retry attempts.  Defaults to 3.
+   */
+  maxRetries?: number;
+
+  /**
+   * The maximum delay in milliseconds allowed before retrying an operation.
+   */
+  maxRetryDelayInMs?: number;
+}