+++ sdk/core/core-rest-pipeline/src/nodeHttpClient.ts
@@ -106,83 +106,58 @@ class NodeHttpClient implements HttpClient {
 
     let responseStream: NodeJS.ReadableStream | undefined;
     try {
-      const result = await new Promise<PipelineResponse>((resolve, reject) => {
-        if (body && request.onUploadProgress) {
-          const onUploadProgress = request.onUploadProgress;
-          const uploadReportStream = new ReportTransform(onUploadProgress);
-          uploadReportStream.on("error", reject);
-          if (isReadableStream(body)) {
-            body.pipe(uploadReportStream);
-          } else {
-            uploadReportStream.end(body);
-          }
-
-          body = uploadReportStream;
-        }
-        const req = this.makeRequest(request, async (res) => {
-          const headers = getResponseHeaders(res);
-
-          const status = res.statusCode ?? 0;
-          const response: PipelineResponse = {
-            status,
-            headers,
-            request
-          };
-
-          // Responses to HEAD must not have a body.
-          // If they do return a body, that body must be ignored.
-          if (request.method === "HEAD") {
-            res.destroy();
-            resolve(response);
-            return;
-          }
-
-          responseStream = shouldDecompress ? getDecodedResponseStream(res, headers) : res;
-
-          const onDownloadProgress = request.onDownloadProgress;
-          if (onDownloadProgress) {
-            const downloadReportStream = new ReportTransform(onDownloadProgress);
-            downloadReportStream.on("error", reject);
-            responseStream.pipe(downloadReportStream);
-            responseStream = downloadReportStream;
-          }
-
-          if (request.streamResponseStatusCodes?.has(response.status)) {
-            response.readableStreamBody = responseStream;
-          } else {
-            try {
-              response.bodyAsText = await streamToText(responseStream);
-            } catch (e) {
-              reject(e);
-            }
-          }
-
-          resolve(response);
+      if (body && request.onUploadProgress) {
+        const onUploadProgress = request.onUploadProgress;
+        const uploadReportStream = new ReportTransform(onUploadProgress);
+        uploadReportStream.on("error", (e) => {
+          logger.error("Error in upload progress", e);
         });
-        req.on("error", (err) => {
-          reject(new RestError(err.message, { code: RestError.REQUEST_SEND_ERROR, request }));
-        });
-        abortController.signal.addEventListener("abort", () => {
-          req.abort();
-          reject(new AbortError("The operation was aborted."));
-        });
-        if (body && isReadableStream(body)) {
-          body.pipe(req);
-        } else if (body) {
-          if (typeof body === "string" || Buffer.isBuffer(body)) {
-            req.end(body);
-          } else if (isArrayBuffer(body)) {
-            req.end(ArrayBuffer.isView(body) ? Buffer.from(body.buffer) : Buffer.from(body));
-          } else {
-            logger.error("Unrecognized body type", body);
-            throw new RestError("Unrecognized body type");
-          }
+        if (isReadableStream(body)) {
+          body.pipe(uploadReportStream);
         } else {
-          // streams don't like "undefined" being passed as data
-          req.end();
+          uploadReportStream.end(body);
         }
-      });
-      return result;
+
+        body = uploadReportStream;
+      }
+
+      const res = await this.makeRequest(request, abortController, body);
+
+      const headers = getResponseHeaders(res);
+
+      const status = res.statusCode ?? 0;
+      const response: PipelineResponse = {
+        status,
+        headers,
+        request
+      };
+
+      // Responses to HEAD must not have a body.
+      // If they do return a body, that body must be ignored.
+      if (request.method === "HEAD") {
+        res.destroy();
+        return response;
+      }
+
+      responseStream = shouldDecompress ? getDecodedResponseStream(res, headers) : res;
+
+      const onDownloadProgress = request.onDownloadProgress;
+      if (onDownloadProgress) {
+        const downloadReportStream = new ReportTransform(onDownloadProgress);
+        downloadReportStream.on("error", (e) => {
+          logger.error("Error in download progress", e);
+        });
+        responseStream.pipe(downloadReportStream);
+        responseStream = downloadReportStream;
+      }
+
+      if (request.streamResponseStatusCodes?.has(response.status)) {
+        response.readableStreamBody = responseStream;
+      } else {
+        response.bodyAsText = await streamToText(responseStream);
+      }
+
+      return response;
     } finally {
       // clean up event listener
       if (request.abortSignal && abortListener) {
@@ -197,8 +172,10 @@ class NodeHttpClient implements HttpClient {
 
         Promise.all([uploadStreamDone, downloadStreamDone])
           .then(() => {
-            request.abortSignal?.removeEventListener("abort", abortListener!);
-            return;
+            // eslint-disable-next-line promise/always-return
+            if (abortListener) {
+              request.abortSignal?.removeEventListener("abort", abortListener);
+            }
           })
           .catch((e) => {
             logger.warning("Error when cleaning up abortListener on httpRequest", e);
@@ -209,8 +186,9 @@ class NodeHttpClient implements HttpClient {
 
   private makeRequest(
     request: PipelineRequest,
-    callback: (res: http.IncomingMessage) => void
-  ): http.ClientRequest {
+    abortController: AbortController,
+    body?: RequestBodyType
+  ): Promise<http.IncomingMessage> {
     const url = new URL(request.url);
 
     const isInsecure = url.protocol !== "https:";
@@ -228,11 +206,36 @@ class NodeHttpClient implements HttpClient {
       method: request.method,
       headers: request.headers.toJSON()
     };
-    if (isInsecure) {
-      return http.request(options, callback);
-    } else {
-      return https.request(options, callback);
-    }
+
+    return new Promise<http.IncomingMessage>((resolve, reject) => {
+      const req = isInsecure ? http.request(options) : https.request(options);
+
+      req.once("response", (res: http.IncomingMessage) => {
+        resolve(res);
+      });
+      req.once("error", (err) => {
+        reject(new RestError(err.message, { code: RestError.REQUEST_SEND_ERROR, request }));
+      });
+      abortController.signal.addEventListener("abort", () => {
+        req.abort();
+        reject(new AbortError("The operation was aborted."));
+      });
+      if (body && isReadableStream(body)) {
+        body.pipe(req);
+      } else if (body) {
+        if (typeof body === "string" || Buffer.isBuffer(body)) {
+          req.end(body);
+        } else if (isArrayBuffer(body)) {
+          req.end(ArrayBuffer.isView(body) ? Buffer.from(body.buffer) : Buffer.from(body));
+        } else {
+          logger.error("Unrecognized body type", body);
+          throw new RestError("Unrecognized body type");
+        }
+      } else {
+        // streams don't like "undefined" being passed as data
+        req.end();
+      }
+    });
   }
 
   private getOrCreateAgent(request: PipelineRequest, isInsecure: boolean): http.Agent {