+++ sdk/identity/identity/test/msalTestUtils.ts
@@ -9,14 +9,9 @@ import {
   pluginForIdentitySDK
 } from "@azure/test-utils-recorder";
 import Sinon, { createSandbox } from "sinon";
-import { assert } from "chai";
-import {
-  OperationTracingOptions,
-  setSpan,
-  setTracer,
-  context as otContext
-} from "@azure/core-tracing";
-import { TestTracer, SpanGraph } from "@azure/test-utils";
+import assert from "assert";
+import { OperationTracingOptions, setSpan, context as otContext } from "@azure/core-tracing";
+import { SpanGraph, setTracer } from "@azure/test-utils";
 import { MsalBaseUtilities } from "../src/msal/utils";
 import * as dotenv from "dotenv";
 dotenv.config();
@@ -173,8 +168,7 @@ export interface TestTracingOptions {
 export function testTracing(options: TestTracingOptions): () => Promise<void> {
   return async function() {
     const { test, children } = options;
-    const tracer = new TestTracer();
-    setTracer(tracer);
+    const tracer = setTracer();
     const rootSpan = tracer.startSpan("root");
 
     const tracingContext = setSpan(otContext.active(), rootSpan);