+++ sdk/storage/storage-file-share/test/utils/index.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { TokenCredential } from "@azure/core-http";
 import { randomBytes } from "crypto";
 import * as fs from "fs";
 import * as path from "path";
@@ -16,7 +17,8 @@ import { StorageSharedKeyCredential } from "../../src/credentials/StorageSharedK
 import { newPipeline } from "../../src/Pipeline";
 import { ShareServiceClient } from "../../src/ShareServiceClient";
 import { extractConnectionStringParts } from "../../src/utils/utils.common";
-import { getUniqueName } from "./testutils.common";
+import { getUniqueName, SimpleTokenCredential } from "./testutils.common";
+import { BlobServiceClient } from "@azure/storage-blob";
 
 export * from "./testutils.common";
 
@@ -46,6 +48,10 @@ export function getGenericBSU(
   });
   const filePrimaryURL = `https://${accountName}${accountNameSuffix}.file.core.windows.net/`;
   return new ShareServiceClient(filePrimaryURL, pipeline);
+}
+
+export function getBlobServceClient(): BlobServiceClient {
+  return BlobServiceClient.fromConnectionString(getConnectionStringFromEnvironment());
 }
 
 export function getBSU(): ShareServiceClient {
@@ -189,6 +195,19 @@ async function streamToBuffer(readableStream: NodeJS.ReadableStream): Promise<Bu
     });
     readableStream.on("error", reject);
   });
+}
+
+export function getTokenCredential(): TokenCredential {
+  const accountTokenEnvVar = `ACCOUNT_TOKEN`;
+  let accountToken: string | undefined;
+
+  accountToken = process.env[accountTokenEnvVar];
+
+  if (!accountToken || accountToken === "") {
+    throw new Error(`${accountTokenEnvVar} environment variables not specified.`);
+  }
+
+  return new SimpleTokenCredential(accountToken);
 }
 
 /**