+++ sdk/cosmosdb/cosmos/CHANGELOG.md
@@ -1,14 +1,16 @@
 # Release History
 
-## 3.12.1 (Unreleased)
+## 3.12.2 (2021-07-21)
 
 ### Features Added
 
-### Breaking Changes
+- Adopted target ES2017 to reduce bundle size.
+
+## 3.12.1 (2021-07-16)
 
 ### Bugs Fixed
 
-### Other Changes
+- Returned default retryPolicy option `fixedRetryIntervalInMilliseconds` to its original default 0.
 
 ## 3.12.0 (2021-07-06)